engine project as only reference from the application
