+++ src/Application/Pages/Index.razor
@@ -1,8 +1,6 @@
 ﻿@page "/"
 @using Microsoft.AspNetCore.SignalR.Client
 @inject NavigationManager Navigation
-@inject TimerTriggerService TimerService
-@inject FeatureService FeatureService
 
 <h1>Willkommen bei Kurmann Videoschnitt</h1>
 
@@ -53,8 +51,7 @@
 
     private async void ExecuteMetadataProcessing()
     {
-        // todo: Implement metadata processing with FeatureService
-        await FeatureService.ProcessMetadataAsync();
+        // todo: send message to server
 
         logs.Add("Metadaten-Verarbeitung gestartet.");
         StateHasChanged();
@@ -62,14 +59,14 @@
 
     private void StartTimerService()
     {
-        TimerService.StartTimer();
+        // todo: send message to server
         logs.Add("Timer-Service gestartet.");
         StateHasChanged();
     }
 
     private void StopTimerService()
     {
-        TimerService.StopTimer();
+        // todo: send message to server
         logs.Add("Timer-Service gestoppt.");
         StateHasChanged();
     }
