+++ src/Application/_Imports.razor
@@ -1,4 +1,3 @@
 ﻿@using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.JSInterop
-@using Kurmann.Videoschnitt.Application.Services
