+++ src/Engine/Engine.csproj
@@ -50,5 +50,9 @@
     <None Include="..\..\LICENSE" Pack="true" PackagePath="\" />
     <None Include="..\..\PackageIcon.png" Pack="true" PackagePath="\" />
   </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Features\MetadataProcessor\MetadataProcessor.csproj" />
+  </ItemGroup>
   
 </Project>
