+++ src/Features/MetadataProcessor/MetadataProcessingService.cs
@@ -8,12 +8,6 @@ public class MetadataProcessingService
     private readonly ILogger<MetadataProcessingService> _logger;
     private readonly DirectoryInfo _metadataProcessingDirectory;
 
-    public event EventHandler<MetadataProcessingEventArgs> MetadataProcessingEvent;
-    protected virtual void OnMetadataProcessingEvent(MetadataProcessingEventArgs e)
-    {
-        MetadataProcessingEvent?.Invoke(this, e);
-    }
-
     public MetadataProcessingService(ILogger<MetadataProcessingService> logger)
     {
         _logger = logger;
@@ -32,7 +26,9 @@ public class MetadataProcessingService
             return Task.CompletedTask;
         }
         _logger.LogInformation("Verzeichnis für Metadatenverarbeitung: {directory}", _metadataProcessingDirectory.FullName);
-        OnMetadataProcessingEvent(new MetadataProcessingEventArgs("Metadatenverarbeitung gestartet auf Verzeichnis " + _metadataProcessingDirectory.FullName));
+        
+        // todo: integrate message service
+        // OnMetadataProcessingEvent(new MetadataProcessingEventArgs("Metadatenverarbeitung gestartet auf Verzeichnis " + _metadataProcessingDirectory.FullName));
 
         _logger.LogInformation("Metadatenverarbeitung abgeschlossen.");
         return Task.CompletedTask;
