+++ src/Features/MetadataProcessor/MetadataProcessor.csproj
@@ -4,8 +4,8 @@
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <RootNamespace>Kurmann.Videoschnitt.Features.MetadatProcessor</RootNamespace>
-    <AssemblyName>Kurmann.Videoschnitt.Features.MetadatProcessor</AssemblyName>
+    <RootNamespace>Kurmann.Videoschnitt.Features.MetadataProcessor</RootNamespace>
+    <AssemblyName>Kurmann.Videoschnitt.Features.MetadataProcessor</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
